[2/2] store: Adapt _copy_move_path to _copy_move
As _copy_move no longer automatically tries to add .gpg to the end of
the path names this now has to be done here.

Also we now try to remove any empty directories before we call git to
remove them from it's index, so that git does not complain about any
uncommmited changes.